Basic auth from env, disable basic auth on staging + dev

+++ pages/api/remove-grid.ts
@@ -1,14 +1,11 @@
 import {NextApiRequest, NextApiResponse} from 'next'
 import {ValidationError} from 'yup'
-import {allowAccessFor} from '../../utils/auth'
+import {ensureAuthentication} from '../../utils/auth'
 import {client} from '../../utils/gql'
 
 export default async (req: NextApiRequest, res: NextApiResponse) => {
-  if (!allowAccessFor(req.headers.authorization, ['kancelaria'])) {
-    res.statusCode = 401
-    res.setHeader('WWW-Authenticate', 'Basic')
-    res.end('Unauthorized')
-  }
+  if (!ensureAuthentication(req, res)) return
+
   try {
     await client.DeleteGridMutation({id: req.body.id})
     res.status(201).end()
